use mailbox system to allow for parallel rechunking
