+++ bin/rechunker
@@ -1,6 +1,8 @@
 #!/usr/bin/env python
 import os.path
 import argparse
+import typing
+
 import strax
 import pandas as pd
 
@@ -13,7 +15,7 @@ def parse_args():
         '--source',
         type=str,
         help="Target directory to rechunk, should be a folder in a "
-             "strax.DataDrictory (one datatype)")
+             "strax.DataDirectory (one datatype)")
     parser.add_argument(
         '--dest', '--destination',
         default=None,
@@ -44,12 +46,26 @@ def parse_args():
         type=str,
         default=None,
         help="Write some information to this file (csv format)")
+    parser.add_argument(
+        '--parallel',
+        type=str,
+        default='False',
+        choices=['False', 'True', 'thread', 'process'],
+        help="Parallelize using threadpool or processpool")
+    parser.add_argument(
+        '--max_workers',
+        type=int,
+        default=4,
+        help="Max workers if parallel is specified")
+    parser.add_argument(
+        '--profile_memory',
+        action='store_true',
+        help="Profile memory usage")
     args = parser.parse_args()
     return args
 
 
-def main():
-    args = parse_args()
+def main(args):
     source_mb = strax.utils.dir_size_mb(args.source)
     report = strax.rechunker(source_directory=args.source,
                              dest_directory=args.dest,
@@ -57,6 +73,8 @@ def main():
                              compressor=args.compressor,
                              target_size_mb=args.target_size_mb,
                              rechunk=args.rechunk,
+                             parallel={'False': False, 'True': True}.get(args.parallel, args.parallel),
+                             max_workers=args.max_workers,
                              )
     if args.dest is not None:
         recompressed_mb = strax.utils.dir_size_mb(args.dest)
@@ -82,4 +100,14 @@ def main():
 
 
 if __name__ == '__main__':
-    main()
+    args = parse_args()
+    if args.profile_memory:
+        from memory_profiler import memory_usage
+        import time
+        start = time.time()
+        mem = memory_usage(proc=(main, (args,)))
+        print(f"Memory profiler says peak RAM usage was: {max(mem):.1f} MB")
+        print(f'Took {time.time() - start:.1f} s = {(time.time() - start) / 3600:.2f} h ')
+        print('Bye, bye')
+    else:
+        main(args)
