+++ strax/__init__.py
@@ -11,6 +11,7 @@ from strax.processing.general import *
 
 from .storage.common import *
 from .storage.files import *
+from .storage.file_rechunker import *
 from .storage.mongo import *
 from .storage.zipfiles import *
