+++ strax/storage/files.py
@@ -1,16 +1,12 @@
 import glob
 import json
-import tempfile
 import os
 import os.path as osp
-import typing
-import time
 from bson import json_util
 import shutil
 
 import strax
 from .common import StorageFrontend
-
 export, __all__ = strax.exporter()
 
 RUN_METADATA_PATTERN = '%s-metadata.json'
@@ -215,6 +211,18 @@ class FileSytemBackend(strax.StorageBackend):
     Metadata is stored in a file called metadata.json.
     """
 
+    def __init__(self, *args, set_target_chunk_mb: int = None, **kwargs, ):
+        """
+        Add set_chunk_size_mb to strax.StorageBackend to allow changing
+        the chunk.target_size_mb returned from the loader, any args or kwargs
+        are passed to the strax.StorageBackend
+
+        :param set_target_chunk_mb: Prior to returning the loaders' chunks,
+            return the chunk with an updated target size
+        """
+        super().__init__(*args, **kwargs)
+        self.set_chunk_size_mb = set_target_chunk_mb
+
     def _get_metadata(self, dirname):
         prefix = dirname_to_prefix(dirname)
         metadata_json = f'{prefix}-metadata.json'
@@ -236,6 +244,11 @@ class FileSytemBackend(strax.StorageBackend):
 
         with open(md_path, mode='r') as f:
             return json.loads(f.read())
+    def _read_and_format_chunk(self, *args, **kwargs):
+        chunk = super()._read_and_format_chunk(*args, **kwargs)
+        if self.set_chunk_size_mb:
+            chunk.target_size_mb = self.set_chunk_size_mb
+        return chunk
 
     def _read_chunk(self, dirname, chunk_info, dtype, compressor):
         fn = osp.join(dirname, chunk_info['filename'])
@@ -366,115 +379,3 @@ class FileSaver(strax.Saver):
 class InvalidFolderNameFormat(Exception):
     pass
 
-
-@export
-def rechunker(source_directory:str,
-              dest_directory: typing.Optional[str] = None,
-              replace: bool = False,
-              compressor: typing.Optional[str] = None,
-              target_size_mb: typing.Optional[str] = None,
-              rechunk: bool = True,
-              progress_bar: bool = True,
-              )-> dict:
-    """
-    Rechunk/Recompress a strax-datatype saved in a FileSystemBackend
-    outside of a strax.Context. For a user-friendly context centered
-    alternative, see strax.Context.copy_to_frontend:
-    github.com/AxFoundation/strax/blob/a0d51fdd3bea52c228c8f74c614fc77bb7cf1bc5/strax/context.py#L1792  # noqa
-
-    One can either specify a destination directory where to store a new
-    copy of this data with <dest_directory> or replace the input file
-    with it's rechunked version.
-
-    This function can either:
-     - rechunk (if <rechunk? is True), probably incrementing the
-        <target_size_mb> is also useful (create larger chunks)
-     - recompress (if a <compressor> is specified)
-
-    One can also rechunk and recompress simultaneously
-
-    :param source_directory: Path to a folder containing a single
-        strax.DataType.
-    :param dest_directory: Head of a folder whereto write new data. If
-        nothing is specified, write to a temporary directory.
-    :param replace: Delete the source_directory and replace it with it's
-        rechunked version
-    :param compressor: Compressor to be used in saving the rechunked
-        data
-    :param target_size_mb: Target size of chunks (uncompressed). As long
-        as a chunk is smaller than this many MB, keep adding new MBs
-        until the chunk is at least target_size_mb or we run out of
-        chunks.
-    :param rechunk: Do we want to rechunk?
-    :param progress_bar: display progress bar
-    :return: Dictionary with some information on the write/load times
-        involved.
-    """
-    if not os.path.exists(source_directory):
-        raise FileNotFoundError(f'No file at {source_directory}')
-    if not replace and dest_directory is None:
-        raise ValueError(f'Specify a destination path <dest_file> when '
-                         f'not replacing the original path')
-    backend_key = os.path.basename(os.path.normpath(source_directory))
-
-    if dest_directory is None and replace:
-        _temp_dir = tempfile.TemporaryDirectory()
-        dest_directory = _temp_dir.name
-    else:
-        _temp_dir = False
-
-
-    if os.path.basename(os.path.normpath(dest_directory)) != backend_key:
-        # New key should be correct! If there is not an exact match,
-        # we want to make sure that we append the backend_key correctly
-        print(f'Will write to {dest_directory} and make sub-folder {backend_key}')
-        dest_directory = os.path.join(dest_directory, backend_key)
-    backend = strax.FileSytemBackend()
-    meta_data = backend.get_metadata(source_directory)
-    source_compressor = meta_data['compressor']
-
-    if compressor is not None:
-        meta_data['compressor'] = compressor
-    if target_size_mb is not None:
-        meta_data['chunk_target_size_mb'] = target_size_mb
-
-    data_loader = backend.loader(source_directory)
-    pbar = strax.utils.tqdm(total = len(meta_data['chunks']), disable=not progress_bar)
-    load_time_seconds = []
-    def loader():
-        """Wrapped loader for bookkeeping load time"""
-        while True:
-            try:
-                t0 = time.time()
-                data = next(data_loader)
-                # Update target chunk size for re-chunking
-                data.target_size_mb = meta_data['chunk_target_size_mb']
-                load_time_seconds.append(time.time()-t0)
-                pbar.n += 1
-                pbar.display()
-            except StopIteration:
-                pbar.close()
-                return
-            yield data
-    print(f'Rechunking {source_directory} to {dest_directory}')
-    saver = backend._saver(dest_directory, metadata=meta_data)
-
-    write_time_start = time.time()
-    saver.save_from(loader(), rechunk=rechunk)
-    load_time = sum(load_time_seconds)
-    write_time = time.time() - write_time_start - load_time
-
-    if replace:
-        print(f'move {dest_directory} to {source_directory}')
-        shutil.rmtree(source_directory)
-        shutil.move(dest_directory, source_directory)
-    if _temp_dir:
-        _temp_dir.cleanup()
-
-    return dict(backend_key=backend_key,
-                load_time=load_time,
-                write_time=write_time,
-                uncompressed_mb= sum([x['nbytes'] for x in meta_data['chunks']]) / 1e6,
-                source_compressor=source_compressor,
-                dest_compressor=meta_data['compressor'],
-                )
